refactor(#758): scope the replay wake poke per service instance
Replace the module-level sendReplayWake singleton with a
makeSendReplayWake factory that closes over its own oRPC client cache
and coalesce state, built once per service in createActivityService and
threaded through the router into stopActivity, trackActivityProgress,
and getAvatarProgression. Per-instance state lets a test inject a zero
coalesce window, so wake-delivery assertions no longer sleep out a
shared one-second window; the client-handle holder and its test-only
reset setter are gone.
