+++ services/activity/src/handlers/get-avatar-progression.ts
@@ -2,7 +2,11 @@ import type { DB } from '@vers/db';
 import type { Kysely } from 'kysely';
 import { parseTerminalCheckpointXP } from '../parse-terminal-checkpoint-xp';
 import type { MissingSessionPayload } from '../types';
-import { sendReplayWake } from '../wake/send-replay-wake';
+
+interface GetAvatarProgressionDeps {
+  readonly db: Kysely<DB>;
+  readonly sendReplayWake: () => void;
+}
 
 /**
  * oRPC handler opts for the authed `getAvatarProgression` procedure.
@@ -36,14 +40,14 @@ interface AvatarProgression {
  * when the avatar doesn't exist or isn't owned by the acting user.
  */
 export async function getAvatarProgression(
-  db: Kysely<DB>,
+  deps: GetAvatarProgressionDeps,
   opts: GetAvatarProgressionOpts,
 ): Promise<AvatarProgression | null> {
   if (opts.context.actingUserId === null) {
     throw opts.errors.UNAUTHORIZED({ data: { reason: 'missing-session' } });
   }
 
-  const rows = await db
+  const rows = await deps.db
     .selectFrom('avatars')
     .leftJoin('activities', (join) =>
       join
@@ -71,7 +75,7 @@ export async function getAvatarProgression(
   const pending = rows.flatMap((row) => findPendingEntry(row));
 
   if (pending.length > 0) {
-    sendReplayWake();
+    deps.sendReplayWake();
   }
 
   return { level: settled.level, pending, xp: settled.xp };
